:zap: Slightly improve performance when sending messages
