+++ src/features/Chat/component/PromptInput/index.ts
@@ -1,0 +1,1 @@
+export * from './PromptInput';
